Fix uninitialized variable warning in wxMSW wxFileSystemWatcher code.
This warning was harmless as the variable was in fact always initialized if
the code where it was used was reached but g++ 4.6.1 is not smart enough to
see it, just as MSVC for which we already had a workaround. So initialize it
explicitly just to suppress the warning.

See #14459.