Use a single taskbar icon for all notifications in wxMSW.
Allocating a new icon for every notification could result in showing many
identical icons in the taskbar notification area if several notification
messages were generated which looked like a bug to the user. It was also
inconsistent with the behaviour in the case when UseTaskBarIcon() was called.

Always behave as in the latter case now, i.e. any subsequent notification
replaces the previous one instead of being shown in addition to it.