Don't assume that wxTE_RICH(2) implies wxTE_MULTILINE for wxTextCtrl in wxOSX.
Under MSW single line text controls can still use these styles, so don't
assume that the control should be made multi-line just because one of them is
specified.

This fixes the appearance of the in-place text control in wxGridCellTextEditor
which is created with wxTE_RICH2 style to work around some wxMSW problem.

Closes #11681.